# gen/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Example food api to learn golang
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: food-api
  version: "1.0"
servers:
- url: //localhost:8080/
paths:
  /food:
    get:
      description: Get all food
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/application.FoodDTO'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad request
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Not found
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Internal Server Error
      summary: Get all food
      tags:
      - food
      x-accepts: application/json
    post:
      description: Create new food item
      responses:
        "201":
          content:
            application/json:
              schema:
                type: string
          description: Created
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad request
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Not found
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Internal Server Error
      summary: Post food
      tags:
      - food
      x-accepts: application/json
  /food/{id}/bite:
    post:
      description: Bite a selected food
      parameters:
      - description: Food ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: OK
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Bad request
        "404":
          content:
            application/json:
              schema:
                type: string
          description: Not found
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Internal Server Error
      summary: Post food bite
      tags:
      - food
      x-accepts: application/json
components:
  schemas:
    application.FoodDTO:
      example:
        price: price
        name: name
        weight: 0
        id: 0
      properties:
        id:
          minimum: 0
          type: integer
        name:
          type: string
        price:
          type: string
        weight:
          minimum: 0
          type: integer
      required:
      - id
      - name
      - price
      - weight
      type: object
x-original-swagger-version: "2.0"